+++ apps/ledger-box/src/modules/wallets/wallet-settings-activity/index.ts
@@ -1,0 +1,1 @@
+export { WalletSettingsActivity } from '#/modules/wallets/wallet-settings-activity/wallet-settings-activity';